Relocate project index module into plugin tree
Moved src/shared/project-index/index.js to src/plugin/src/project-index/index.js because the project index exclusively supports plugin identifier-case features. Updated imports, docs, and tests so everything continues to reference the module from its new home.
