+++ docs/examples/identifier-case/locals-first.prettierrc.mjs
@@ -2,7 +2,7 @@ import path from "node:path";
 import { mkdir, readFile, writeFile } from "node:fs/promises";
 import { fileURLToPath } from "node:url";
 
-import { buildProjectIndex } from "./node_modules/root/src/shared/project-index/index.js";
+import { buildProjectIndex } from "./node_modules/root/src/plugin/src/project-index/index.js";
 
 const configFilePath = fileURLToPath(import.meta.url);
 const projectRoot = path.dirname(configFilePath);
