+++ src/plugin/tests/identifier-case-asset-conflicts.test.js
@@ -4,7 +4,7 @@ import os from "node:os";
 import path from "node:path";
 import { describe, it } from "node:test";
 
-import { buildProjectIndex } from "../../shared/project-index/index.js";
+import { buildProjectIndex } from "../src/project-index/index.js";
 import { prepareIdentifierCasePlan } from "../src/identifier-case/local-plan.js";
 
 async function createCollisionProject() {
